performance: mark REPL and Doc code as non-specializeable
Implicit return is bad for compiler performance (and sometimes runtime
performance) and can adversely affect code readability,
so every function which does _not_ return a value should end in a `return` statement.

Here, we also introduce a new meaning to the `@nospecialize` macro, and
a new macro `@specialize` to reverse its effect.
When used without arguments, it applies to all arguments of the parent.
In local scope, this means the containing function.
In global scope, this means all methods subsequently defined.